Print the versioned module path to stdout
Update the main function to print the version and the path of the packed
module to stdout.  The versioned module path can be used in a go.mod
require directive to ensure that the cached version will be used.

Implement the Stringer interface for CachedModule, that returns the
versioned module path.